pyproject.toml
[project]
name = "photo_to_excel"
version = "1.0.0"
authors = [
    {name="Julia Marcinkowska", email="juliamarcinkowska98@gmail.com" },
]
maintainers = [
    {name="Julia Marcinkowska", email="juliamarcinkowska98@gmail.com" },
]
description = "Covert your photos to Excel file"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: MacOS",
]
license = { file = "LICENSE" }

[project.urls]
Homepage = "https://github.com/juliam98/photo_to_excel"
Issues = "https://github.com/juliam98/photo_to_excel/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src/"]
include = ["photo_to_excel"]

[tool.poetry.dependencies]
python = "^3.13"
openpyxl = "^3.1.5"
pillow = "^11.1.0"
